+++ atompack/crystal/__init__.py
@@ -1,6 +1,6 @@
 """Abstractions for generating and modifying atomic structures with long range order."""
 
-from atompack.crystal.components import Basis, LatticeParameters, LatticeVectors
+from atompack.crystal.components import (Basis, LatticeParameters, LatticeVectors)
 from atompack.crystal.crystal import Crystal, UnitCell
 from atompack.crystal.spatial import MillerIndex, Orientation, Plane
 from atompack.crystal.transform import Transform